refactor(blocks): extract block-related constants into block-constants.js
Move MINIMUMDOCKDISTANCE, LONGSTACK, SPATIAL_GRID_CELL_SIZE, CAMERAVALUE,
VIDEOVALUE, NOTEBLOCKS, PITCHBLOCKS, and ALLOWED_CONNECTIONS out of
blocks.js into a dedicated block-constants.js module, following the
existing logoconstants.js extraction pattern. blocks.js keeps referencing
them as globals; loader.js loads block-constants before blocks so the
values are available in both the browser and CommonJS/Jest environments.
No behavior or public API changes.
