+++ js/blocks.js
@@ -25,7 +25,9 @@
    getVoiceSynthName, i18nSolfege, last, MathUtility, mixedNumber,
    piemenuBlockContext, prepareMacroExports, ProtoBlock,
     setOctaveRatio, splitScaleDegree, splitSolfege, updateTemperaments,
-    docById, define, BlocksDependencies, deepClone, pubsub
+    docById, define, BlocksDependencies, deepClone, pubsub,
+    MINIMUMDOCKDISTANCE, LONGSTACK, SPATIAL_GRID_CELL_SIZE,
+    CAMERAVALUE, VIDEOVALUE, NOTEBLOCKS, PITCHBLOCKS, ALLOWED_CONNECTIONS
 */
 
 /* global showZoomOverlay */
@@ -34,6 +36,9 @@
    Global locations
    - js/activity.js
         createjs
+   - js/block-constants.js
+        MINIMUMDOCKDISTANCE, LONGSTACK, SPATIAL_GRID_CELL_SIZE,
+        CAMERAVALUE, VIDEOVALUE, NOTEBLOCKS, PITCHBLOCKS, ALLOWED_CONNECTIONS
    - js/block.js
         Block
    - js/piemenus.js
@@ -52,21 +57,7 @@
         setOctaveRatio, splitScaleDegree, splitSolfege,
         updateTemperaments
 */
-/**
- * Minimum distance (squared) between two docks required before
- * connecting them.
- */
-const MINIMUMDOCKDISTANCE = 400;
-
-/** Soft limit on the number of blocks in a single stack. */
-const LONGSTACK = 300;
-
-/**
- * Spatial grid cell size in pixels for O(1) nearest-dock lookups.
- * Chosen so that MINIMUMDOCKDISTANCE (20px radius at default scale)
- * is always covered by checking a block's cell plus its 8 neighbors.
- */
-const SPATIAL_GRID_CELL_SIZE = 50;
+// Constants moved to js/block-constants.js
 
 /**
  * Lazy-initialized Sets for O(1) collapsible type checks in hot paths.
@@ -85,84 +76,6 @@ function getInlineCollapsiblesSet() {
     if (!_inlineCollapsiblesSet) _inlineCollapsiblesSet = new Set(INLINECOLLAPSIBLES);
     return _inlineCollapsiblesSet;
 }
-
-/** Special value flags to uniquely identify these media blocks. */
-const CAMERAVALUE = "##__CAMERA__##";
-const VIDEOVALUE = "##__VIDEO__##";
-
-const NOTEBLOCKS = ["newnote", "osctime"];
-const PITCHBLOCKS = ["pitch", "steppitch", "hertz", "pitchnumber", "nthmodalpitch", "playdrum"];
-
-const ALLOWED_CONNECTIONS = new Set([
-    "vspaceout:vspacein",
-    "vspacein:vspaceout",
-    "in:out",
-    "out:in",
-    "in:vspaceout",
-    "vspaceout:in",
-    "out:vspacein",
-    "vspacein:out",
-    "numberin:numberout",
-    "numberin:anyout",
-    "numberout:numberin",
-    "anyout:numberin",
-    "textin:textout",
-    "textin:anyout",
-    "textout:textin",
-    "anyout:textin",
-    "booleanout:booleanin",
-    "booleanin:booleanout",
-    "mediain:mediaout",
-    "mediaout:mediain",
-    "mediain:textout",
-    "textout:mediain",
-    "filein:fileout",
-    "fileout:filein",
-    "casein:caseout",
-    "caseout:casein",
-    "vspaceout:casein",
-    "casein:vspaceout",
-    "vspacein:caseout",
-    "caseout:vspacein",
-    "solfegein:anyout",
-    "solfegein:solfegeout",
-    "solfegein:textout",
-    "solfegein:noteout",
-    "solfegein:scaledegreeout",
-    "solfegein:numberout",
-    "anyout:solfegein",
-    "solfegeout:solfegein",
-    "textout:solfegein",
-    "noteout:solfegein",
-    "scaledegreeout:solfegein",
-    "numberout:solfegein",
-    "notein:solfegeout",
-    "notein:scaledegreeout",
-    "notein:textout",
-    "notein:noteout",
-    "solfegeout:notein",
-    "scaledegreeout:notein",
-    "textout:notein",
-    "noteout:notein",
-    "pitchout:anyin",
-    "gridout:anyin",
-    "anyin:textout",
-    "anyin:mediaout",
-    "anyin:numberout",
-    "anyin:anyout",
-    "anyin:fileout",
-    "anyin:solfegeout",
-    "anyin:scaledegreeout",
-    "anyin:noteout",
-    "textout:anyin",
-    "mediaout:anyin",
-    "numberout:anyin",
-    "anyout:anyin",
-    "fileout:anyin",
-    "solfegeout:anyin",
-    "scaledegreeout:anyin",
-    "noteout:anyin"
-]);
 
 /**
  * Blocks holds the list of blocks and most of the block-associated
